new controller is added
