+++ WebAplicationTestMVCSOL/WebAplicationTestMVC/Views/Flashcards/AddFlashcard.cshtml
@@ -4,7 +4,7 @@
 }
 
 <h1>Make a Flashcard</h1>
-@using (Html.BeginForm("SubmitNewFlashcard", "Home", FormMethod.Post))
+@using (Html.BeginForm("SubmitNewFlashcard", "Flashcards", FormMethod.Post))
 {
     <label for="question">FlashCard Question:</label>
     <br>
