+++ WebAplicationTestMVCSOL/WebAplicationTestMVC/Views/Home/Index.cshtml
@@ -28,8 +28,8 @@
 
     <div>
         <button id="createStudySet" class="btn btn-info">+ Create Study Set</button>
-        <button id="showFormButton" class="btn btn-info">+Import Flashcard Set</button>
-        <form id="myForm" method="post" enctype="multipart/form-data" asp-controller="Home" asp-action="ImportDB">
+        <button id="showFormButton" class="btn btn-info">+ Import Flashcard Set</button>
+        <form id="myForm" method="post" enctype="multipart/form-data" asp-controller="StudySet" asp-action="ImportDB">
             <label for="file">Upload flashcards:</label>
             <input type="file" name="file" id="file" required accept=".xlsx" />
             <button type="submit" class="btn btn-warning">Import</button>
@@ -42,7 +42,7 @@
         {
             <a class="col-sm- set-box"
                style="background-color: @(studySet.Color.ToString().ToLower());text-decoration: none;"
-               asp-area="" asp-controller="Home" asp-action="StudySets" asp-route-studySetName="@studySet.studySetName">
+               asp-area="" asp-controller="StudySet" asp-action="StudySets" asp-route-studySetName="@studySet.studySetName">
                 <div class="set-name">@studySet.studySetName.Substring(0, studySet.studySetName.Length - 5) *</div>
             </a>
