+++ WebAplicationTestMVCSOL/WebAplicationTestMVC/Views/StudySet/StudySets.cshtml
@@ -33,7 +33,7 @@
 
     <hr class="soften" />
 
-    <a class="btn btn-primary" asp-area="" asp-controller="Home" asp-action="AddFlashcard" asp-route-studySetName="@Model.studySetName">Create flashcards manually</a>
+    <a class="btn btn-primary" asp-area="" asp-controller="Flashcards" asp-action="AddFlashcard" asp-route-studySetName="@Model.studySetName">Create flashcards manually</a>
     <a class="btn btn-primary" data-toggle="modal" data-target="#DataSettings" asp-area="">Learn flashcards</a>
